requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
commonmark==0.9.1
coverage==6.4.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.6.0
gunicorn==20.1.0
heroku==0.1.4
httpie==3.2.1
idna==3.3
multidict==6.0.2
Pillow==9.2.0
platformdirs==2.5.1
psycopg2==2.9.3
pygame==2.1.2
Pygments==2.12.0
PyJWT==2.4.0
PySocks==1.7.1
python-dateutil==1.5
python-decouple==3.6
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
views==0.3
virtualenv==20.14.0
whitenoise==6.2.0
